move non-ui code to utils
