--- utils.py
@@ -1,5 +1,0 @@
-import sys
-
-
-def eprint(*args, **kwargs):
-    print(*args, file=sys.stderr, **kwargs)
