+++ utils/eprint.py
@@ -1,0 +1,4 @@
+import sys
+
+def eprint(*args, **kwargs):
+    print(*args, file=sys.stderr, **kwargs)
